requirements.txt
Django
django-cors-headers
djangorestframework
python-dotenv
coverage